move original download options to own component
